% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LBCF_IS}
\alias{LBCF_IS}
\title{Parallel Logit-BART-IS for ITE estimation with prediction intervals}
\usage{
LBCF_IS(
  lambda_mu,
  lambda_tau,
  num_models,
  num_trees_mu,
  num_trees_tau,
  seed,
  ytrain,
  original_datamat,
  ztrain,
  pihat_train,
  beta_par,
  test_datamat,
  test_pihat,
  ncores,
  outsamppreds,
  nu,
  a_mu,
  a_tau,
  lambdaBCF,
  valid_trees,
  tree_prior,
  imp_sampler,
  alpha_BCF_mu,
  beta_BCF_mu,
  alpha_BCF_tau,
  beta_BCF_tau,
  include_pi2,
  fast_approx,
  PIT_propensity,
  lower_prob,
  upper_prob,
  root_alg_precision,
  maxit,
  eps_f,
  eps_g,
  num_iter,
  include_cate_intervals
)
}
\arguments{
\item{seed}{The seed for random number generation.}

\item{original_datamat}{The original test data. This matrix must have the same number of columns (variables) as the training data. Currently all variables must be continuous. The test data does not need to be transformed before being entered to this function.}

\item{beta_par}{The power to which the likelihood is to be raised. For BMA, set beta_par=1.}

\item{ncores}{The number of cores to be used in parallelization.}

\item{maxit}{Maximum number of iterations for the quasi-Newton algorithm that finds the MAP estimate for each model (required for Laplace approximation).}

\item{eps_f}{Parameter for MAP algorithm stopping criterion. Iteration stops if |f-f'|/|f|<eps_f, where f and f' are the current and previous value of the objective function (negative log likelihood) respectively.}

\item{eps_g}{Parameter for MAP algorithm stopping criterion. Iteration stops if ||g|| < eps_g * max(1, ||beta||), where beta is the current coefficient vector and g is the gradient.}

\item{lambda}{A real number between 0 and 1 that determines the splitting probability in the prior (which is used as the importance sampler of tree models). Quadrianto and Ghahramani (2015) recommend a value less than 0.5 .}

\item{num_trees}{The number of trees to be sampled.}

\item{num_cats}{The number of possible values for the outcome variable.}

\item{y}{The training data vector of outcomes. This must be a vector of integers between 1 and num_cats.}

\item{alpha_parameters}{Vector of prior parameters.}
}
\value{
A List containing 1. A vector of predictions, and 2. A matrix of prediction intervals, the first row corresponds to the lower quantile, the second row is the median, and the third row is the upper quantile.
}
\description{
A parallelized implementation of Logit Bayesian Additive Regression Trees using importance sampling of models.
}
